Split animations out of battle scene
